requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
python-jose
redis
celery
python-dateutil